Lift ILLocalVariable and ILSequencePoint into Internal.IL
