+++ src/ILCompiler.Compiler/src/Compiler/SymbolReader/PdbSymbolReader.cs
@@ -5,43 +5,10 @@
 using System;
 using System.Collections.Generic;
 
+using Internal.IL;
+
 namespace ILCompiler.SymbolReader
 {
-    /// <summary>
-    ///  IL sequence point record
-    /// </summary>
-    public struct ILSequencePoint
-    {
-        public readonly int Offset;
-        public readonly string Document;
-        public readonly int LineNumber;
-        // TODO: The remaining info
-
-        public ILSequencePoint(int offset, string document, int lineNumber)
-        {
-            Offset = offset;
-            Document = document;
-            LineNumber = lineNumber;
-        }
-    }
-
-    /// <summary>
-    ///  IL local variable debug record
-    /// </summary>
-    public struct ILLocalVariable
-    {
-        public readonly int Slot;
-        public readonly string Name;
-        public readonly bool CompilerGenerated;
-
-        public ILLocalVariable(int slot, string name, bool compilerGenerated)
-        {
-            Slot = slot;
-            Name = name;
-            CompilerGenerated = compilerGenerated;
-        }
-    }
-
     /// <summary>
     /// Abstraction for reading Pdb files
     /// </summary>
